- Made all the wxSpinCtrls on the TEXTUREx editor smaller horizontally
git-svn-id: http://mancubus.net/svn/hosted/slade/trunk@435 1ddb354d-3e0c-0410-833a-d70222086f97

+++ src/GLTexture.cpp
@@ -303,9 +303,9 @@ bool GLTexture::draw2d(int32_t x, int32_t y) {
 		glTranslated(x, y, 0);
 
 		size_t tex_index = 0;
-		size_t top = 0;
+		double top = 0;
 		while (top < height) {
-			size_t left = 0;
+			double left = 0;
 			while (left < width) {
 				// Bind the texture
 				glBindTexture(GL_TEXTURE_2D, tex[tex_index].id);
